+++ package.json
@@ -17,8 +17,8 @@
     "husky": "^9.1.5",
     "lucide-react": "^0.428.0",
     "next": "14.2.5",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7"
   },
