+++ src/app/layout.tsx
@@ -1,25 +1,32 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "../styles/globals.css";
+"use client"
+import { ReactLenis } from '@studio-freight/react-lenis';
+import { Inter } from 'next/font/google';
+import React from 'react';
 
-const inter = Inter({ subsets: ["latin"] });
+import '../styles/globals.css';
 
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
+const inter = Inter({ subsets: ['latin'] });
+
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
-}>) 
-{
+}>) {
+  const lenisOptions = {
+    lerp: 5,
+    duration: 3,
+    smoothTouch: false,
+    smooth: true,
+  };
+
   return (
-    <html lang="en">
-      <body className={`${inter.className} overflow-x-hidden w-screen`}>
-        {children}
+    <ReactLenis root options={lenisOptions}>
+      <html lang="en">
+        <body className={`${inter.className} overflow-x-hidden w-screen`}>
+          {children}
         </body>
       </html>
-);
+    </ReactLenis>
+  );
 }
