+++ src/app/page.tsx
@@ -1,11 +1,8 @@
 import Events from '@/components/sections/Events';
-import SmoothScrolling from "@/components/smoothscrolling";
 export default function Home() {
   return (
-    <SmoothScrolling>
       <main>
         <Events/>
       </main>
-    </SmoothScrolling>  
   );
 }
